honeycomb: move search-specific honey to separate package to reduce imports in honey package
